Fix: 백엔드 - 결제 주문테이블과 연동
- 주문,매장,점주,고객 번호 가져오기
- 주문1개에 결제1개만 생기도록

+++ backend/src/main/java/com/pudding/base/domain/pay/repository/PayRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PayRepository extends JpaRepository<Pay, Long>{
+public interface PayRepository extends JpaRepository<Pay, Integer>{
 
 }
